+++ usr/src/cmd/cmd-inet/usr.sbin/snoop/snoop_filter.c
@@ -23,6 +23,9 @@
  * Use is subject to license terms.
  * Copyright 2012 Milan Jurik. All rights reserved.
  */
+/*
+ * Copyright 2017 Hayashi Naoyuki
+ */
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -592,7 +595,7 @@ want_packet(uchar_t *pkt, int len, int origlen)
 			/*
 			 * Handle 3 possible alignments
 			 */
-			switch ((((unsigned)base) + off) % sizeof (uint_t)) {
+			switch ((((uintptr_t)base) + off) % sizeof (uint_t)) {
 			case 0:
 				*sp = *(uint_t *)(base + off);
 				break;
