+++ usr/src/cmd/fs.d/nfs/statd/sm_proc.c
@@ -23,6 +23,9 @@
 /*
  * Copyright 2015 Nexenta Systems, Inc.  All rights reserved.
  */
+/*
+ * Copyright 2017 Hayashi Naoyuki
+ */
 
 /*
  * Copyright 2005 Sun Microsystems, Inc.  All rights reserved.
@@ -442,7 +445,7 @@ insert_mon(mon *monp)
 
 	if (debug)
 		(void) printf("add_mon(%x) %s (id=%d)\n",
-		    (int)new, new->id.mon_id.mon_name, *((int *)new->id.priv));
+		    (int)(intptr_t)new, new->id.mon_id.mon_name, *((int *)new->id.priv));
 
 	/*
 	 * Record the name, and all addresses which have been registered
@@ -602,7 +605,7 @@ delete_onemon(char *mon_name, my_id *my_idp, mon_entry **monitor_q)
 				/* found */
 				if (debug)
 					(void) printf("delete_mon(%x): %s\n",
-					    (int)nl, mon_name ?
+					    (int)(intptr_t)nl, mon_name ?
 					    mon_name : "<NULL>");
 				/*
 				 * Remove the monitor name from the
@@ -861,7 +864,7 @@ pr_mon(char *name)
 			(void) printf("*****monitor_q:\n ");
 			while (nl != NULL) {
 				(void) printf("%s:(%x), ",
-				    nl->id.mon_id.mon_name, (int)nl);
+				    nl->id.mon_id.mon_name, (int)(intptr_t)nl);
 				nl = nl->nxt;
 			}
 			mutex_unlock(&mon_table[hash].lock);
@@ -877,7 +880,7 @@ pr_mon(char *name)
 			(void) printf("*****monitor_q:\n ");
 			while (nl != NULL) {
 				(void) printf("%s:(%x), ",
-				    nl->id.mon_id.mon_name, (int)nl);
+				    nl->id.mon_id.mon_name, (int)(intptr_t)nl);
 				nl = nl->nxt;
 			}
 			(void) printf("\n");
